+++ app/src/main/java/com/gildder/invenbras/gestionactivos/models/CTipoActivo.java
@@ -52,8 +52,8 @@ public class CTipoActivo {
 
 
     public Long Insertar(TipoActivo tipoActivo) {
-        SQLiteDatabase db = dbHelper.getWritableDatabase();
         if (!Existe(String.valueOf(tipoActivo.getId()))){
+            SQLiteDatabase db = dbHelper.getWritableDatabase();
             ContentValues values = new ContentValues();
             values.put(_id,tipoActivo.getId());
             values.put(_tipo, tipoActivo.getTipo());
@@ -63,7 +63,7 @@ public class CTipoActivo {
             Long resultID = db.insert(_Tabla, null, values);
 
             Log.i("INF:", "Se adiciono correctamente el tipo "+ tipoActivo.getId());
-
+            db.close();
             return resultID;
         }else{
             Log.i("INF:", "NO Se adiciono correctamente el tipo "+ tipoActivo.getId());
@@ -76,22 +76,22 @@ public class CTipoActivo {
 
     public Long Actualizar(TipoActivo tipoActivo) {
 
-        if (!Existe((String.valueOf(tipoActivo.getId())))){
+        if (Existe((String.valueOf(tipoActivo.getId())))){
             SQLiteDatabase db = dbHelper.getWritableDatabase();
 
             ContentValues values = new ContentValues();
             values.put(_id,tipoActivo.getId());
             values.put(_tipo, tipoActivo.getTipo());
             values.put(_nombre,tipoActivo.getNombre());
-            values.put(_descripcion,tipoActivo.getDescripcion());
+            values.put(_descripcion, tipoActivo.getDescripcion());
 
             Long resultID = (long) db.update(_Tabla, values, "id=?", new String[]{String.valueOf(tipoActivo.getId())});
 
             Log.i("INF:", "Se adiciono correctamente el tipo " + tipoActivo.getId());
-
+            db.close();
             return resultID;
         }else{
-            Log.i("INF:", "NO Se adiciono correctamente el tipo "+ tipoActivo.getId());
+            Log.e("ERROR:", "No se actualizo correctamente porque el tipo activo No existeo " + tipoActivo.getId());
 
             return (long) -1;
         }
@@ -99,10 +99,17 @@ public class CTipoActivo {
 
     public Long Eliminar(int id)
     {
-        SQLiteDatabase db = dbHelper.getWritableDatabase();
-        Long resultID = (long) db.delete(_Tabla,  _id + "=?", new String[]{String.valueOf(id)});
+        if (Existe((String.valueOf(id)))){
+            SQLiteDatabase db = dbHelper.getWritableDatabase();
+            Long resultID = (long) db.delete(_Tabla, _id + "=?", new String[]{String.valueOf(id)});
+            db.close();
 
-        return resultID;
+            return resultID;
+        }else{
+            Log.e("ERROR:", "No se elimino correctamente porque el tipo activo No existe " + id);
+
+            return (long) -1;
+        }
     }
 
 
@@ -124,6 +131,7 @@ public class CTipoActivo {
         }
 
         cursor.close();
+        db.close();
 
         Log.i("INF:", "Se obtubo el tipo " + id);
         return  tipoActivo;
@@ -151,6 +159,67 @@ public class CTipoActivo {
         }
 
         cursor.close();
+        db.close();
+
+        Log.i("INF:", "Se obtuvo todos los tipos ");
+
+        return  listaTipos;
+    }
+
+    public ArrayList<TipoActivo> GetOneTipo(){
+        ArrayList<TipoActivo> listaTipos = new ArrayList<>();
+
+        SQLiteDatabase db = dbHelper.getReadableDatabase();
+        Cursor cursor = db.query(_Tabla, new String[] {_id, _tipo, _nombre, _descripcion},
+                null, null, null, null, null );
+
+        if (cursor != null){
+            while (cursor.moveToNext()) {
+                TipoActivo tipoActivo = new TipoActivo();
+                tipoActivo.setId(cursor.getInt(0));
+                tipoActivo.setTipo(cursor.getString(1));
+                tipoActivo.setNombre(cursor.getString(2));
+                tipoActivo.setDescripcion(cursor.getString(3));
+
+                if(cursor.isFirst()) {
+                        listaTipos.add(tipoActivo);
+                }else{
+                    if (!tipoActivo.getTipo().toString().equals(listaTipos.get(listaTipos.size()-1).getTipo().toString()))
+                        listaTipos.add(tipoActivo);
+                }
+            }
+        }
+
+        cursor.close();
+        db.close();
+
+        Log.i("INF:", "Se obtuvo todos los tipos ");
+
+        return  listaTipos;
+    }
+
+    public ArrayList<TipoActivo> GetAllTipo(String tipo){
+        ArrayList<TipoActivo> listaTipos = new ArrayList<>();
+
+        SQLiteDatabase db = dbHelper.getReadableDatabase();
+        Cursor cursor = db.query(_Tabla, new String[] {_id, _tipo, _nombre, _descripcion},
+                _tipo + "=?",
+                new String[]{tipo}, null, null, null);
+
+        if (cursor != null){
+            while (cursor.moveToNext()) {
+                TipoActivo tipoActivo = new TipoActivo();
+                tipoActivo.setId(cursor.getInt(0));
+                tipoActivo.setTipo(cursor.getString(1));
+                tipoActivo.setNombre(cursor.getString(2));
+                tipoActivo.setDescripcion(cursor.getString(3));
+
+                listaTipos.add(tipoActivo);
+            }
+        }
+
+        cursor.close();
+        db.close();
 
         Log.i("INF:", "Se obtuvo todos los tipos ");
 
@@ -163,6 +232,7 @@ public class CTipoActivo {
         if(db != null) {
             count = (int) DatabaseUtils.queryNumEntries(db, _Tabla);
         }
+        db.close();
 
         return count;
     }
@@ -170,9 +240,9 @@ public class CTipoActivo {
 
     private boolean Existe(String id) {
         boolean result = false;
-        SQLiteDatabase db = dbHelper.getReadableDatabase();
 
-        if(db != null && Count()>0) {
+        SQLiteDatabase db = dbHelper.getReadableDatabase();
+        if(db != null) {
             Cursor cursor = db.query(_Tabla, new String[]{_id, _tipo, _nombre, _descripcion},
                     _id + "=?",new String[]{id}, null, null, null);
             result = (cursor.getCount() > 0);
@@ -181,6 +251,8 @@ public class CTipoActivo {
 
             Log.i("INF:", "El Tipo " + id + "Ya existe");
         }
+        db.close();
+
         return result;
